refactor: resolve #337 - extract DeviceScreenManager from WebWorkerDeviceAdapter
Extract screen state management (cursor, color/palette, screen sync,
shared display buffer, BG graphics, sprite position tracking) into a
dedicated DeviceScreenManager class. Adapter becomes a thinner facade
at 413 lines (down from 497), well under the 500-line limit.

Closes #337
